Inline tests cases which don't report
* Remove duplicate tests
* Move tests to a Spec context in order to tidy up the class
* Inline a case file
